website/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
faiss-cpu==1.7.4
openai==1.3.5
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.28.1

